requirements.txt
gradio
python-dotenv
google-genai
moviepy
gTTS
pydub
